Add 'decomment.c' program
this program take at least two argument:
- the first the name of the program itself for launch the application
- the second is the name of the file to scan for searching comments

An optional third argument is the name of another file in which the
formatted output, i.e., without comments, is copied.

If the program is launch without the third argument, then
standard output is the target for printing the formatted version
of the file.